(bug) - Honour custom insync in noop mode
Prior to this commit, a custom insync message was not honoured when
applying a manifest in noop mode. This was later tracked down the the
`noop` method found in the resource harness class.

I've updated the method to match that of the `sync` method below,
ensuring that the functionality of the custom isync provider feature is
not any different whether running puppet in noop or no-noop mode.