Update JRuby in tests to 9.4.7.0
This commit updates the version of JRuby used in spec tests from 9.4.3.0
to 9.4.7.0, which matches what is used in the latest version of Puppet
Server 8.x. (See: puppetlabs/jruby-utils@1180711.)

JRuby 9.4.7.0 includes bug fixes to several issues that we have worked
around in Puppet in the past:

- Dir.glob not properly supporting wildcard syntax (worked around in
  Puppet in 663062a, resolved in JRuby in jruby/jruby@a39cd49).
- format and sprintf not supporting %a and %A (worked around in Puppet
  in 334ea05, resolved in JRuby in jruby/jruby#7996).

Because of these bug fixes in JRuby, this commit also removes Puppet's
workarounds.