Bump to JRuby 9.3.14.0, reenable test
This commit updates JRuby to 9.3.14.0 in tests. This version of JRuby
addresses a bug with Dir.glob (see jruby/jruby@5b06409) that we had
previously worked around in both code and tests (see 10d219a). Because
the Dir.glob bug has now been addressed, this commit also removes the
workarounds we had added.

Co-authored-by: Shubham Shinde <shubham.shinde@perforce.com>