(PUP-9997) Pass cwd to execute method
Dir.chdir is problematic because it affects all threads in the current process
and if puppet is started with a current working directory it doesn't have
traverse/execute permission to, then it won't be able to restore the cwd at the
end of the Dir.chdir block.

Puppet supports three execution implementations: posix, windows and stub. The
first two already support the `cwd` option. Puppetserver injects its stub
implementation and it recently added support for `cwd`, see SERVER-3051.