Add pre_resource_eval to POSIX provider
This commit adds API documentation, updates a call to, and adds a spec test for
the pre_resource_eval class function in the POSIX file provider (which was added
in #9349). When called, pre_resource_eval will create a class variable,
selinux_handle, which can be used when handling data for SELinux. Since the
handle is a class variable, we can avoid running into performance issues since
the handle can be re-used instead of needing to make a new one each time.

Additionally, since the old method wasn't completely removed & replaced
(and instead deprecated), less changes to old spec tests are needed as their
calls to the deprecated method aren't impacted.