Check for nil before closing Uniquefile
On Windows, we by default assign stdout for the child process to a file:

    stdout = Puppet::FileSystem::Uniquefile.new('puppet')

However, if TMPDIR, TEMP or TMP env vars refer to a non-existent directory, then
`Uniquefile.new` will raise, leaving `stdout` set to nil, so the later call to
`stdout.close` raised NoMethodError.

So check for nil before closing. We don't have this issue on posix, because
it uses an IO pipe instead of Uniquefile.

Fixes #9385