(PUP-1881) Correct Windows runinterval behavior
Previously when Puppet parsed a runinterval of 0 on Windows, it would
set the runinterval to 1800 seconds, Puppet's default runinterval value.
This was incorrect behavior, as on other platforms and in the
documentation, a runinterval of 0 indicates that Puppet should run
continuously.

This commit updates the Windows daemon to set the runinterval to 1800
seconds when it receives an empty string and cast to integer in all
other cases.