(PUP-12029) Add syslog and getoptlong dependencies
Running puppet with Ruby 3.3.0 prints:

    $ bundle exec puppet --version
    /home/josh/work/puppet/lib/puppet/settings.rb:4: warning: getoptlong was loaded from the standard library, but will no longer be part of the default gems since Ruby 3.4.0. Add getoptlong to your Gemfile or gemspec.
    /home/josh/work/puppet/lib/puppet/util/feature.rb:116: warning: syslog was loaded from the standard library, but will no longer be part of the default gems since Ruby 3.4.0. Add syslog to your Gemfile or gemspec.

Since getoptlong is necessary to parse the command line and is a pure ruby gem,
it's safe to add it as a runtime dependency in puppet.gemspec.

Since syslog is a feature and requires native extensions, add it to the features
group in the Gemfile (excluding Windows and JRuby). It will also need to be
added to the runtime on non-Windows platforms.

The versions were chosen to match what is currently shipped in ruby 3.2.4, but
updates to the patch component are allowed.