Use bundler-cache to bundle install and cache gems
The setup-ruby action makes it trivial to bundle install gems from a
cache and update the cache when the run completes.

When using `bundler-cache`, we can't use `bundle config` commands, but
we can achieve the same effect specifying BUNDLE_SET as an env variable.

See https://github.com/ruby/setup-ruby for details