(PUP-12046) Send facts for `puppet catalog download`
Facts were not always sent with the catalog request when
using the catalog face. This change allows the face to send
facts for the request so that the catalog returned is
compiled with the correct set of facts.

(cherry picked from commit c47de0311f851b8512183efa44b42c5ac04989b0)