(PUP-12065) Add task to generate configuration reference
    bundle exec rake references:configuration

It runs `puppet doc -r configuration` to generate references/configuration.md.

In order for the output to not be dependent on the hostname/domain facts, allow
them to be overridden as an environment variable. This can't be done using
FACTER_*, because structured facts like `networking.hostname` can't be overridden
by an env variable.

To determine if any meaningful changes have been made, you can run:

    git diff  -I built_from_commit -I 'This page was generated from the Puppet source code' references/