Render man/man5/puppetconf.5 correctly
Previously, `rake genmanpages` generated warnings when converting ronn to man:

    roff: ./man/man5/puppetconf.5
    warn: unrecognized block tag: "hr"
    warn: unexpected text: "Return a YAML hash with up to three keys:"

The first warning was because definition lists need to have a space between each
item. As a result the entire line "Return a YAML hash with up to three keys" was
omitted from the output.

The second warning was because '* * *' is not valid in ronn.