Add files to include in tarball
Previously, Puppet relied on the in-house Packaging library to build all
artifacts related to Puppet. Puppet moved away from using Packaging to
build gems in 867ce9c, which also removed the files key used to build
tarballs.

This commit re-adds the files key (as a list instead of a string, as
recommended by Packaging) to enable the packaging:tar Rake task, which
builds source tarballs of Puppet.