Clarify soft limit on number of facts
Puppet warns when the total number of fact values exceeds a soft limit,
as PuppetDB performance can start to degrade as that number grows.

However, the language that Puppet currently uses to talk about this
limit can be ambiguous. We want to emphasize that the total number of
fact values--counting all children of all facts (every value in all
arrays, hashes, etc.)--is what is relevant to performance.

This commit adds more explicit language about what this number is
counting.