Disable reports by default
The default value for this goes all the way back to when it was
introduced in 566bf7820e56b3be503a05645cad54152309b20f. I'd argue it a
bad default, because it just writes files without any clean up process
in place. That only causes hard disks to fill up.

Having reports should be an opt in feature, where they make a careful
choice. PuppetDB is an obvious recommended option but not one that can
be relied on by default.