(PUP-12058) Add task to generate man (as markdown) reference
    bundle exec rake references:man

This task generates an overview page that lists all of the puppet commands.

Next, we run the gen_manpages rake task that generates roff formatted man pages
in the man directory. Then we convert roff to markdown using pandoc. We
unfortunately can't generate markdown directly as legacy apps are hardcoded to
generate ronn, such as when running `puppet agent --help`.

The task requires the pandoc-ruby gem, so add it to the documentation bundler
group. Since the group is optional, check to see if the gem is present before
loading it, similar to how we handle ronn.

The task also requires the pandoc executable.