(PUP-12075) Render puppet man page correctly
Previously, the man/man8/puppet.8 man page wasn't formatted correctly
with each subcommand and summary on a new line. This was because `puppet
--help` doesn't output ronn format, just raw text. So generate
`puppet.8.ronn` in valid ronn format so that it can be converted to
roff. Use `man --local-file man/man8/puppet.8` to verify the results.

Also don't blindly run `<bin>/<app> --help` for every app. That's left
over from when puppet had multiple bin stubs.

Since ronn shells out to groff, abort if the latter is not installed.

Hardcode the http_user_agent value so it's not sensitive to the puppet
version or the ruby version/architecture the user is running.