(PUP-12077) Respect rich_data setting in base context
Since Puppet 6.0 "datafication" has inspected the context for the value
of rich_data. However, in only some code paths does the value in the
context get overridden with a value taken from the settings. This means
in some cases the rich_data value will always be true or always be
false, regardless of how the user has configured the rich_data setting.
And, in the case the simply calling `to_data_hash` on a resource the
rich_data value will always be false.

This updates the base_context to set rich_data to the settings value,
ensuring that the default value for rich_data in the context is the
value users have set.

Additional changes are primarily where tests still assumed the default
value of rich_data was false, with one exception - YAML serialization in
the resource application will break if the internal rich_data `__pcore`
values are output. This forces rich_data to be false for that code path
in the resource application.

Fixes GH #9470