(PUP-12055) Build references on commit to main
When a commit is pushed to main, build all of the references, including man
pages. If the only changes are due to SHA and dates, then do nothing. Otherwise,
commit all of the references and man pages.

The action is only triggered when a commit is pushed to main, but not on pull
requests so that we don't have to worry about untrusted inputs.

The action is only triggered when the repo owner is puppetlabs, so it won't
trigger on forks.

The action uses full SHAs for the pandoc and add-and-commit actions.

If changes are detected, the action creates a commit whose author is GitHub
Actions <actions@github.com> with message:

    Update references [no-promote]

And pushes the commit to the main branch. It uses the repository's GITHUB_TOKEN
to accomplish this. We don't need to worry about recursive workflow runs[1]:

    When you use the repository's GITHUB_TOKEN to perform tasks, events
    triggered by the GITHUB_TOKEN ... will not create a new workflow run. This
    prevents you from accidentally creating recursive workflow runs.

[1] https://docs.github.com/en/actions/writing-workflows/choosing-when-your-workflow-runs/triggering-a-workflow