(PUP-12061) Recalculate splay when splaylimit changes
If puppet.conf was modified, but splaylimit wasn't, then we recalculated
the splay offset. This goes against the desired behavior of only splaying when
the agent is initially started and running every runinterval seconds from that
point forward or until runinterval/splaylimit are changed in the future.

Now the SplayJob remembers its last splaylimit so that it only recalculates
splay if the limit changes. It's important to recalculate splay so that the
agent chooses a slot over the new range, either smaller or bigger than before.