(PUP-12061) Handle changing `splay` in puppet.conf
Previously, enabling or disabling splay in puppet.conf did not work if the
agent was already running periodically.

If the agent was started with splay disabled, then enabling it would cause a
NoMethodError, when trying to call Puppet::Scheduler::Job#splay_limit

If the agent was started with splay enabled, then disabling it would have no
effect, since we never recalculated the splay limit.

To handle these cases, always create a SplayJob for the agent_run job and set
its splay_limit to either the limit or 0, depending on whether splay is enabled
or not. Setting a splay_limit to 0 causes the splay to also be set to 0 because
rand(1) always returns 0.