(PUP-12061) Add test cases for runinterval and splaylimit interaction
Since splaylimit defaults to runinterval, changes to runinterval affect splay.
Add tests covering cases where runinterval is increased and decreased. It's
important to recalculate splay to reduce thundering herds. For example, if
runinterval is changed from 30 mins to 60 mins, we want the compute a new slot
in the 60 min interval to decrease server load.