Handle additional edge cases in puppet.gemspec
Previously, running `gem build puppet.gemspec` on windows included the
windows-specific runtime dependencies in the "ruby" puppet-<version>.gem,
because of the `Gem.win_platform?` condition.

Now add the runtime dependencies if we're running on that platform via bundler.
or if we're building a gem for that platform. In the latter case, it doesn't
matter which platform we're running on.

This also relaxes the darwin and mingw platforms, to future proof against other
variations of darwin and mingw ucrt.